rebuild aliases immediately if they are already available
On pod restart, it takes a small amount of time for the aliases file
to appear - as such, inotifyd tends to trigger the reload script and
everything is normal.

On container restart $VIRTUAL_LOC may already exist - as such we may
have to wait an extended period of time before inotifyd kicks in. We
can mitigate this by doing (most) of what `/reload` does immidiately.